requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.2.1
cffi==1.17.1
cryptography==44.0.1
debugpy==1.8.12
ecdsa==0.19.0
fastapi[standard]==0.115.8
greenlet==3.1.1
idna==3.10
msgpack==1.1.0
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pynvim==0.5.2
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
sqlmodel==0.0.22
starlette==0.45.3
typing_extensions==4.12.2
